feature: take a step further and seperate node types
